+++ ProjetSteam/ProjetSteam/SteamAPIManager.h
@@ -1,0 +1,17 @@
+#pragma once
+#ifndef STEAMAPIMANAGER_H
+#define STEAMAPIMANAGER_H
+#include <steam_api.h>
+
+class SteamAPIManager
+{
+private:
+
+public:
+	SteamAPIManager();
+
+	bool SteamAPIInit();
+	void SteamAPIQuit();
+};
+
+#endif