+++ changelog.md
@@ -1,3 +1,14 @@
+## [0.0.3] - 2022-02-10
+### Ajout
+-Mise en place de l'API Steam.
+-Création de la classe GameManager.
+-Création de la classe SDLManager.
+-Création de la classe SteamAPIManager.
+
+## [0.0.2] - 2022-02-05
+### Ajout
+-Reprise en main de la SDL avec le chargement d'une image.
+
 ## [0.0.1] - 2022-02-03
 ### Ajout
 -Création du projet.